Refactor mixins out into separate files
Splits the mixins related to form elements and the general template helpers into separate files.
