+++ index.js
@@ -1,3 +1,3 @@
 'use strict';
 
-module.exports = require('./lib/template-mixins');
+module.exports = require('./lib');
