+++ lib/template-mixins.js
@@ -1,12 +1,12 @@
 'use strict';
 
 var fs = require('fs'),
-    path = require('path'),
-    querystring = require('querystring');
+    path = require('path');
 
 var Hogan = require('hogan.js'),
-    _ = require('underscore'),
-    moment = require('moment');
+    _ = require('underscore');
+
+var renderer = require('./render');
 
 // This returns a middleware that places mixins against the `res.locals` object.
 //
@@ -16,25 +16,7 @@ var Hogan = require('hogan.js'),
 //   - viewEngine: the type of view, defaults to 'html'.
 //   - sharedTranslationsKey: used to find translations relatively within
 //     the translations.json. Useful for field and button labels.
-module.exports = function (options, deprecated) {
-
-    if (arguments.length === 2) {
-        /*eslint no-console: 0*/
-        console.warn('hof-template-mixins: Passing field config to mixins is deprecated');
-        console.warn('hof-template-mixins: Field config will be loaded from res.locals.options.fields');
-        options = deprecated;
-    }
-
-    options = options || {};
-    options.fields = options.fields || {};
-
-    var viewsDirectory = options.viewsDirectory || path.resolve(__dirname, '../');
-    var viewEngine = options.viewEngine || 'html';
-    var sharedTranslationsKey = options.sharedTranslationsKey || '';
-
-    if (sharedTranslationsKey && !sharedTranslationsKey.match(/\.$/)) {
-        sharedTranslationsKey += '.';
-    }
+module.exports = function (options) {
 
     var PANELMIXIN = 'partials/mixins/panel';
 
@@ -82,7 +64,7 @@ module.exports = function (options, deprecated) {
 
     return function (req, res, next) {
 
-        var roots = [].concat(req.app.get('views')).concat(viewsDirectory);
+        var roots = [].concat(req.app.get('views')).concat(options.viewsDirectory);
         var View = req.app.get('view');
 
         // wrap in try catch to throw an error if any one template cannot be resolved
@@ -91,11 +73,11 @@ module.exports = function (options, deprecated) {
                 if (compiled[relativeTemplatePath]) {
                     return;
                 }
-                var viewExtension = '.' + viewEngine;
+                var viewExtension = '.' + options.viewEngine;
                 var engines = {};
                 engines[viewExtension] = {};
                 var view = new View(relativeTemplatePath, {
-                    defaultEngine: viewEngine,
+                    defaultEngine: options.viewEngine,
                     root: roots,
                     engines: engines
                 });
@@ -109,19 +91,15 @@ module.exports = function (options, deprecated) {
             return next(e);
         }
 
-        var hoganRender = function (text, ctx) {
-            if (!text) { return ''; }
-            ctx = Object.assign({}, res.locals, ctx);
-            return Hogan.compile(text).render(ctx);
-        };
+        var hoganRender = renderer(res);
 
         var t = function (key) {
-            return hoganRender(req.translate(sharedTranslationsKey + key), this);
+            return hoganRender(req.translate(options.sharedTranslationsKey + key), this);
         };
 
         // Like t() but returns null on failed translations
         var conditionalTranslate = function (key) {
-            key = sharedTranslationsKey + key;
+            key = options.sharedTranslationsKey + key;
             var translated = req.translate(key);
             return translated !== key ? translated : null;
         };
@@ -141,10 +119,10 @@ module.exports = function (options, deprecated) {
             var match = child.match(re);
             res.locals.partials = res.locals.partials || {};
             if (match) {
-                return fs.readFileSync(res.locals.partials['partials-' + match[1]] + '.' + viewEngine).toString();
+                return fs.readFileSync(res.locals.partials['partials-' + match[1]] + '.' + options.viewEngine).toString();
             } else if (child === 'html' || res.locals[child]) {
-                var panelPath = path.join(viewsDirectory, PANELMIXIN + '.' + viewEngine);
-                var localsPanelPath = path.join(res.locals.partials['partials-mixins-panel'] + '.' + viewEngine);
+                var panelPath = path.join(options.viewsDirectory, PANELMIXIN + '.' + options.viewEngine);
+                var localsPanelPath = path.join(res.locals.partials['partials-mixins-panel'] + '.' + options.viewEngine);
                 if (res.locals.partials['partials-mixins-panel']) {
                     return fs.readFileSync(localsPanelPath).toString();
                 } else {
@@ -233,7 +211,7 @@ module.exports = function (options, deprecated) {
                     return template.render(Object.assign({
                         renderMixin: renderMixin.bind(this)
                     }, res.locals, this), _.mapObject(res.locals.partials, function (path) {
-                        return Hogan.compile(fs.readFileSync(path + '.' + viewEngine).toString());
+                        return Hogan.compile(fs.readFileSync(path + '.' + options.viewEngine).toString());
                     }));
                 }
             };
@@ -475,91 +453,6 @@ module.exports = function (options, deprecated) {
             res.locals[name] = handler;
         });
 
-        res.locals.currency = function () {
-            return function (txt) {
-                txt = hoganRender(txt, this);
-                var value = parseFloat(txt);
-                if (isNaN(value)) {
-                    return txt;
-                } else if (value % 1 === 0) {
-                    value = value.toString();
-                } else {
-                    value = value.toFixed(2);
-                }
-                return '£' + value;
-            };
-        };
-
-        res.locals.date = function () {
-            return function (txt) {
-                txt = (txt || '').split('|');
-                var value = hoganRender(txt[0], this);
-                return moment(value).format(txt[1] || 'D MMMM YYYY');
-            };
-        };
-
-        res.locals.hyphenate = function () {
-            return function (txt) {
-                var value = hoganRender(txt, this);
-                return value.trim().toLowerCase().replace(/\s+/g, '-');
-            };
-        };
-
-        res.locals.uppercase = function () {
-            return function (txt) {
-                return hoganRender(txt, this).toUpperCase();
-            };
-        };
-
-        res.locals.lowercase = function () {
-            return function (txt) {
-                return hoganRender(txt, this).toLowerCase();
-            };
-        };
-
-        res.locals.selected = function () {
-            return function (txt) {
-                var bits = txt.split('='),
-                    val;
-                if (this.values && this.values[bits[0]] !== undefined) {
-                    val = this.values[bits[0]].toString();
-                }
-                return val === bits[1] ? ' checked="checked"' : '';
-            };
-        };
-
-        /**
-        * Use on whole sentences
-        */
-        res.locals.time = function () {
-            return function (txt) {
-                txt = hoganRender(txt, this);
-                txt = txt.replace(/12:00am/i, 'midnight').replace(/^midnight/, 'Midnight');
-                txt = txt.replace(/12:00pm/i, 'midday').replace(/^midday/, 'Midday');
-                return txt;
-            };
-        };
-
-        res.locals.t = function () {
-            return function (txt) {
-                txt = hoganRender(txt, this);
-                return t.apply(req, [txt, this]);
-            };
-        };
-
-        res.locals.url = function () {
-            return function (url) {
-                url = hoganRender(url, this);
-                return req.baseUrl ? path.resolve(req.baseUrl, url) : url;
-            };
-        };
-
-        res.locals.qs = function () {
-            return function (query) {
-                return '?' + querystring.stringify(Object.assign({}, req.query, querystring.parse(query)));
-            };
-        };
-
         res.locals.renderField = function () {
             return function (key) {
                 if (key) {
@@ -585,6 +478,7 @@ module.exports = function (options, deprecated) {
                 }
             };
         };
+
         next();
     };
