+++ package.json
@@ -39,6 +39,7 @@
   "devDependencies": {
     "chai": "^3.5.0",
     "eslint": "^2.10.2",
+    "eslint-config-homeoffice": "^2.1.2",
     "express": "^4.15.2",
     "istanbul": "^0.4.3",
     "mocha": "^2.4.5",
